feat: Seperate view and logic invitationPopUp
